refactor(table): extract the operation vocabulary into its own module
Operation, UpdateMode, the two group types, RewriteGroup, RewrittenIndex and
UpdatedFragmentOffsets are the data model the rest of the module is written
against. Separating the definitions from the code that applies them leaves each
side readable on its own: this file answers what a transaction can say, and
manifest_build answers what happens when it is applied.
